Remove BioPhysicalInteraction
This class is just a container for one particular sbml tag, it juste store xml bunk from sbml in a string alongside a metaId. This could be easily replaced by a map in the import class. It also have method for parsing xml creating map and formating map content in xml. this should be moved to an util class in the import package. Another method is used for getting gene by parsing annotation. This should be in the import class (possible duplicated code here). Remove ref to the bioAnnotations in other biodata. Use of this class in import class have just been commented for now.