Remove BioPathwayStep
Unused class. work in progress. Suposed to enable iteration over reactions from a pathway. Implies linearity, which is not guraranted by the BioPathway class, otherwise it become difficult to predict comportment. Methods from the BioGraph package should be more useable for the kind of task this is supposed to solve