Remove getNewId methods from BioNetwork
Relying on Palsson convention, only used in exchange reaction addition to generate new id. Changed in order to be more flexible, the new ids are passed as parameter and managed outside this class. The Orphan manager have been changed so it generate itself new ids for exchange reaction producing orphan .