Add method for subgraph extraction in GraphFactory
Extract sub-graph of a main graph given a list of vertex. The obtained sub-graph contains all the edges from the main graph linking vertices from the given list. The obtained sub-graph is not a deep copy of the main graph, any changes to the shared edges attributes will be effective in both graphs.
Another method was added, allowing to Create a graph from a list of edges and a list of vertices. Vertex connected by the edges from the list should be present in the given vertex collection.